Fix up the user profile showcase
